requirements.txt
pytest>=3.7
delegator.py==0.1.0
gitpython==2.1
